Move classes to Internal folder/namespace
